+++ src/lib/ui/events.ts
@@ -16,6 +16,13 @@ import type {MainNavView} from './ui';
 export interface Dispatch {
 	(eventName: 'log_in', params: LoginRequest): Promise<ApiResult<{session: ClientAccountSession}>>;
 	(eventName: 'log_out', params: void): Promise<ApiResult<void>>;
+	(eventName: 'toggle_main_nav', params: void): void;
+	(eventName: 'toggle_secondary_nav', params: void): void;
+	(eventName: 'set_main_nav_view', params: MainNavView): void;
+	(eventName: 'set_mobile', params: boolean): void;
+	(eventName: 'select_persona', params: {persona_id: number}): void;
+	(eventName: 'select_community', params: {community_id: number | null}): void;
+	(eventName: 'select_space', params: {community_id: number; space_id: number}): void;
 	(eventName: 'create_community', params: {name: string; persona_id: number}): Promise<
 		ApiResult<{community: Community}>
 	>;
@@ -45,13 +52,6 @@ export interface Dispatch {
 	): Promise<ApiResult<{file: File}>>;
 	(eventName: 'read_files', params: {space_id: number}): Promise<ApiResult<{files: File[]}>>;
 	(eventName: 'query_files', params: {space_id: number}): Readable<Readable<File>[]>;
-	(eventName: 'toggle_main_nav', params: void): void;
-	(eventName: 'toggle_secondary_nav', params: void): void;
-	(eventName: 'set_main_nav_view', params: MainNavView): void;
-	(eventName: 'set_mobile', params: boolean): void;
-	(eventName: 'select_persona', params: {persona_id: number}): void;
-	(eventName: 'select_community', params: {community_id: number | null}): void;
-	(eventName: 'select_space', params: {community_id: number; space_id: number}): void;
 }
 
 export interface UiHandlers {
@@ -59,6 +59,13 @@ export interface UiHandlers {
 		ctx: DispatchContext<LoginRequest, ApiResult<{session: ClientAccountSession}>>,
 	) => Promise<ApiResult<{session: ClientAccountSession}>>;
 	log_out: (ctx: DispatchContext<void, ApiResult<void>>) => Promise<ApiResult<void>>;
+	toggle_main_nav: (ctx: DispatchContext<void, void>) => void;
+	toggle_secondary_nav: (ctx: DispatchContext<void, void>) => void;
+	set_main_nav_view: (ctx: DispatchContext<MainNavView, void>) => void;
+	set_mobile: (ctx: DispatchContext<boolean, void>) => void;
+	select_persona: (ctx: DispatchContext<{persona_id: number}, void>) => void;
+	select_community: (ctx: DispatchContext<{community_id: number | null}, void>) => void;
+	select_space: (ctx: DispatchContext<{community_id: number; space_id: number}, void>) => void;
 	create_community: (
 		ctx: DispatchContext<{name: string; persona_id: number}, ApiResult<{community: Community}>>,
 	) => Promise<ApiResult<{community: Community}>>;
@@ -97,13 +104,6 @@ export interface UiHandlers {
 		ctx: DispatchContext<{space_id: number}, ApiResult<{files: File[]}>>,
 	) => Promise<ApiResult<{files: File[]}>>;
 	query_files: (ctx: DispatchContext<{space_id: number}, void>) => Readable<Readable<File>[]>;
-	toggle_main_nav: (ctx: DispatchContext<void, void>) => void;
-	toggle_secondary_nav: (ctx: DispatchContext<void, void>) => void;
-	set_main_nav_view: (ctx: DispatchContext<MainNavView, void>) => void;
-	set_mobile: (ctx: DispatchContext<boolean, void>) => void;
-	select_persona: (ctx: DispatchContext<{persona_id: number}, void>) => void;
-	select_community: (ctx: DispatchContext<{community_id: number | null}, void>) => void;
-	select_space: (ctx: DispatchContext<{community_id: number; space_id: number}, void>) => void;
 }
 
 // generated by src/lib/ui/events.gen.ts
